==> crawling/1.select_parks_by_review_score.ipynb <==
import pandas as pd

# csv 파일 불러오기
df = pd.read_csv("서울시 관광특구 주요공원 리스트.csv")

# 컬럼 이름 변경
df = df.rename(columns={"구": "gu", "공원명": "park_name", "네이버맵 기준 블로그 리뷰 수": "navermap_reviews_count", "카카오맵 기준 블로그 리뷰 수": "kakaomap_reviews_count", "리뷰 수 총합": "total_reviews"})

# 관광 특구만 뽑아내는 작업 필요
df = df[df["gu"].isin(["중구", "용산구", "종로구", "송파구", "강남구", "마포구"])]

# 용산구 별도 저장
yongsan_df = df[df["gu"] == "용산구"]

# 용산구 외 데이터
other_df = df[df["gu"] != "용산구"]

df.head()

df.info()

# 구 별로 total_review를 기준으로 정렬
df_sorted = df.sort_values(["gu", "total_reviews"], ascending=[True, False])

df.head()

# 구별 Q3 이상 공원만 선택할 수 있게 하는 함수
def select_q3(group):
    q3 = group["total_reviews"].quantile(0.75)
    return group[group["total_reviews"] >= q3]

# 전체 기준: 상위 15% 이상인 공원 추출
q85 = df["total_reviews"].quantile(0.85)
highly_reviewed_df = df[df["total_reviews"] >= q85]

highly_reviewed_df

# 구별 Q3 기준 + 전체 상위 15% 병합
q3_selection = other_df.groupby("gu", group_keys=False).apply(select_q3)

# 중복 제거 + 병합
final_df = pd.concat([q3_selection, yongsan_df, highly_reviewed_df], ignore_index=True).drop_duplicates()

# 저장
final_df.to_csv("selected_parks.csv", index=False)

# 확인
print(f"최종 선택된 공원 수: {len(final_df)}개")
final_df
