==> crawling/2.select_search_query.ipynb <==
import pandas as pd

### 장소 리스트 정리

parks_df = pd.read_csv("selected_parks.csv")
parks_df = parks_df[["park_name"]]
parks_df["category"] = "주요공원"
parks_df

han_river_df = pd.DataFrame({
    "park_name": ['강서한강공원', '난지한강공원', '양화한강공원', '망원한강공원', '여의도한강공원', '이촌한강공원', 
                   '반포한강공원', '잠원한강공원', '뚝섬한강공원', '잠실한강공원', '광나루한강공원'],
    "category": "한강공원"
})
han_river_df

places_df = pd.concat([parks_df, han_river_df], ignore_index=True)
places_df

### 키워드 접미사 리스트

# suffix_keywords = {
#     "기본": ["후기", "리뷰"],
#     "활동": ["산책", "나들이"],
#     "풍경/뷰": ["풍경", "뷰"],
#     "계절_봄": ["봄", "벚꽃", "꽃", "개화"],
#     "계절_여름": ["여름", "녹음", "푸름", "초록"],
#     "계절_가을": ["가을", "단풍", "은행", "낙엽", "억새"],
#     "계절_겨울": ["겨울", "눈", "설경"],
#     "시간대": ["노을", "석양", "야경", "일출", "일몰", "해돋이", "해넘이"],
#     "감성_긍정": ["힐링"],
#     "감성_부정": ["혼잡"],
#     "접근성": ["가는 길", "가는 법"]
# }	

# 너무 길어서... 키워드 접미사 간소화
selected_keywords = ["후기", "리뷰", "산책", "풍경", "힐링", "단풍", "벚꽃"]

# 평평한 키워드 리스트로 변경
# flat_suffix_keywords = []
# for category, words in selected_keywords.items():
#     flat_suffix_keywords.extend(words)

### 검색어 조합 생성

search_queries = []
for _, row in places_df.iterrows():
    for suffix in selected_keywords:
        search_queries.append({
            "search_query": f"{row['park_name']} {suffix}",
            "park_name": row["park_name"],
            "category": row["category"],  # 그대로 유지 가능
            "suffix": suffix
        })

search_df = pd.DataFrame(search_queries)
search_df

### 저장

search_df.to_csv("search_query_list.csv", index=False)
