==> crawling/4.filter_naver_blog_reviews.ipynb <==
import pandas as pd

# 크롤링된 데이터 불러오기
df = pd.read_csv("naver_blog_reviews.csv")
df.shape  # 행 수와 열 수 확인
df.head()  # 데이터 미리 보기


def contains_any_keyword(text, keyword_list):
    return any(keyword in text for keyword in keyword_list)


# 예시 키워드 (원하는 대로 수정 가능)
# irrelevant_keywords = [
#     "맛집", "카페", "레스토랑", "식당", "커피", "디저트",
#     "임대", "부동산", "사무실", "월세", "상가",
#     "속눈썹", "왁싱", "피부", "의료", "클리닉",
#     "매물", "술집", "뷰티", "네일", "병원", "피부과",
#     "헬스장", "헬스", "헬스클럽", "스파", "마사지",
#     "분양", "아파트", "오피스텔", "분양권",
#     "옷", "가방", "광고", "협찬", "이벤트", "체험단", "배송", "제품", 
#     "리뷰어", "쇼핑몰", "구매", "브랜드", "패션", "화장품", 
#     "뷰티", "스킨케어", "마스크팩", "택배", "제품후기", "할인"
# ]
strict_irrelevant_keywords = [
    "임대", "부동산", "사무실", "월세", "상가",
    "왁싱", "의료", "클리닉",
    "분양", "아파트", "오피스텔", "분양권",
    "택배", "쇼핑몰", "브랜드", "리뷰어", "체험단"
]


# 관련 없는 행 제거
filtered_df = df[~df["content"].apply(lambda x: contains_any_keyword(str(x), strict_irrelevant_keywords))]

# 결과 확인
print(f"필터링 전: {len(df)}개 → 필터링 후: {len(filtered_df)}개")


filtered_df.to_csv("naver_blog_reviews_filtered.csv", index=False)
print("필터링된 결과 저장 완료: naver_blog_reviews_filtered.csv")


df = pd.read_csv("naver_blog_reviews_filtered.csv")
park_counts = df["park_name"].value_counts()

# print(park_counts.describe())
print(park_counts.head(10))  # 상위 10개 공원 확인
print(park_counts.tail(10))  # 상위 10개 공원 확인
